Move HasAnswerFilter and all its deps to form app
Fixes #530

As the answer filter is specific to form it needs to be in that app
to avoid circular dependencies.
